+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 - cosmwasm-std: Implement `Decimal{,256}::checked_mul` and
   `Decimal{,256}::checked_pow`.
 - cosmwasm-std: Implement `Sub`/`SubAssign` for `Uint64`.
+- cosmwasm-std: Implement `Mul`/`MulAssign` for `Uint64`.
 - cosmwasm-crypto: Upgrade ed25519-zebra to version 3.
 - cosmwasm-vm: Upgrade Wasmer to 2.2.1